// src/modules/client-adm/usecase/find-client/find-client.usecase.dto.ts
export interface FindClientInputDTO {
    id: string
}

export interface FindClientOutputDTO {
    id: string
    name: string
    document: string
    email: string
    street: string
    number: string
    complement: string
    city: string
    state: string
    zipCode: string
    createdAt: Date
    updatedAt: Date
}